Rename DataTableConfig component
