+++ javascript/app-framework/src/components/DataTableConfig/DataTableConfig.utils.ts
