+++ javascript/app-framework/src/components/DataTableConfig/Parser/Parser.styles.ts
