+++ javascript/app-framework/src/components/DataTableConfig/Parser/Parser.types.ts
