+++ javascript/app-framework/src/components/DataTableConfig/Parser/index.ts
