+++ javascript/app-framework/src/components/DataTableConfig/Shape/Shape.tsx
