+++ javascript/app-framework/src/components/DataTableConfig/Shape/TableLayoutOptions.utils.tsx
